+++ src/ffi/marshal.rs
@@ -2,9 +2,9 @@ use super::PyObject;
 use std::os::raw::{c_char, c_int};
 
 extern "C" {
-    #[cfg_attr(PyPy, link_name = "PyMarshal_WriteObjectToString")]
+    #[cfg_attr(PyPy, link_name = "PyPyMarshal_WriteObjectToString")]
     pub fn PyMarshal_WriteObjectToString(object: *mut PyObject, version: c_int) -> *mut PyObject;
 
-    #[cfg_attr(PyPy, link_name = "PyMarshal_ReadObjectFromString")]
+    #[cfg_attr(PyPy, link_name = "PyPyMarshal_ReadObjectFromString")]
     pub fn PyMarshal_ReadObjectFromString(data: *const c_char, len: isize) -> *mut PyObject;
 }